requirements.txt
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
distlib==0.3.7
dnspython==2.4.2
filelock==3.13.1
flasgger==0.9.7.1
Flask==3.0.0
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
pipenv==2023.11.15
platformdirs==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.14.1
six==1.16.0
virtualenv==20.25.0
Werkzeug==3.0.1
